Add UpnpServiceFactory
 - Since Spring's Service class has become bloated, UPnP instantiation
   will be moved to a factory class.
